Move Chain implementation to chain.rs
